Move classes to dubbo-cluster
