+++ dubbo-cluster/src/main/java/org/apache/dubbo/rpc/cluster/loadbalance/statistics/CpuUsageServiceImpl.java
@@ -1,10 +1,9 @@
-package org.apache.dubbo.monitor.dubbo;
+package org.apache.dubbo.rpc.cluster.loadbalance.statistics;
 
 import org.apache.dubbo.common.utils.NamedThreadFactory;
 import org.apache.dubbo.rpc.RpcContext;
 
 import com.alibaba.metrics.Gauge;
-import com.alibaba.metrics.ManualClock;
 import com.alibaba.metrics.MetricName;
 import com.alibaba.metrics.os.linux.CpuUsageGaugeSet;
